+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change log
 
+## [0.4.7](https://github.com/ColdMagi/tjscope/pull/9) - 2022-08-21
+
+### Added
+
+- scrollable tabs in _comments_ module
+- scrollable tabs in _posts_ module
+- post likers rating tab (by user) like in _comments_ module ([#1](https://github.com/ColdMagi/tjscope/issues/1))
+
+### Changed
+
+- color scheme of tabs in _main_ to `cyan`
+- color scheme of tabs in _comments_ module to `teal`
+- color scheme of tabs in _entries_ module to `orange`
+- variant of tabs in _comments_ module from `enclosed` to `soft-rounded`
+- variant of tabs in _entries_ module from `enclosed` to `soft-rounded`
+
 ## [0.4.5] - 2022-08-20
 
 ### Added